requirements.txt
transformers
torch
pillow
numpy